New clientlib ver

+++ src/main/java/com/kaltura/client/types/Asset.java
@@ -109,7 +109,7 @@ public class Asset extends ObjectBase {
 	/**
 	 * The size (in Bytes) of the asset
 	 */
-	private Integer sizeInBytes;
+	private Long sizeInBytes;
 
 	// id:
 	public String getId(){
@@ -208,7 +208,7 @@ public class Asset extends ObjectBase {
 	}
 
 	// sizeInBytes:
-	public Integer getSizeInBytes(){
+	public Long getSizeInBytes(){
 		return this.sizeInBytes;
 	}
 
@@ -236,7 +236,7 @@ public class Asset extends ObjectBase {
 		partnerData = GsonParser.parseString(jsonObject.get("partnerData"));
 		partnerDescription = GsonParser.parseString(jsonObject.get("partnerDescription"));
 		actualSourceAssetParamsIds = GsonParser.parseString(jsonObject.get("actualSourceAssetParamsIds"));
-		sizeInBytes = GsonParser.parseInt(jsonObject.get("sizeInBytes"));
+		sizeInBytes = GsonParser.parseLong(jsonObject.get("sizeInBytes"));
 
 	}
